requirements.txt
requests == 2.24.0
mysql-connector-python == 8.0.22
